refactor(property): split property controller and fix room availability in search
Split the previous monolithic property controller into separate create/query/manage controllers and updated routes to use them. Fixed search availability logic in apps/api/src/services/property-search.service.ts so bookings and room availabilities are checked per room under Rooms.some.AND. Minor frontend tweak: removed delete confirmation in tenant properties list. No DB migrations required. Please add/update tests for search availability and price sorting.
